V 9 -> Implemented: Map View Full Functioning (Only UI Changes necessary) Event Details have going button working with DB and Views Information Correctly. Discover Sorts by Lit or Nah (Home.Swift) Needs to streamline code
