+++ Events/CustomPinAnnotation.swift
@@ -1,0 +1,17 @@
+//
+//  CustomPinAnnotation.swift
+//  Events
+//
+//  Created by Omar Mihilmy on 1/31/16.
+//  Copyright © 2016 FV iMAGINATION. All rights reserved.
+//
+
+import MapKit
+import Parse
+import UIKit
+
+class CustomPinAnnotation: MKPointAnnotation{
+    
+    var EventObject: PFObject!
+
+}
